refactor(bazel): Move all supported rules to the separate directory
